+++ CHANGELOG.md
@@ -22,8 +22,8 @@ and this project adheres to [Semantic Versioning].
   - added `__eq__` method
   - added `__str__` method
   - added `__repr__` method
-  - added `__add__` method
-  - added `__sub__` method
+  - added `__add__` and `__radd__` method
+  - added `__sub__` and `__rsub__` method
   - added `__mul__` and `__rmul__` method
   - added `conjugate` method
   - added `inverse` method